Refined 'Contributary'
To share source code with the jsh project, code must be GPL compatible. Compliance with the Hypersoft Contributory License in a source distribution affirms this.

GPL'd Code may not be contributed to jsh, because it cannot be licensed to a contributary. A developer who has written GPL'd code, that he/she would like to contribute to our project should consider requesting permission to offer such code under a  'Hypersoft Contributary' license. Otherwise, such additions cannot be accepted by this project.

The licensing of this software system, is such to provide a license compatible gateway for an interoperable software system, that is otherwise interoperable in operation. Participation in development of a project of such scope suggests that you would like to create GPL compatible code. Use of a project of such scope suggests that you would like to provide a compatible option, upgrade or supplement with an existing software system, using a basis of this software.

The edits to this document should completely reflect this.